Re-add history actions to mainwindow
History actions are back in main-window to allow shortcut usage and make
the navigation toolbar work again. Clicking the action will move you back
or forward in history in the current active viewspace. Similarly, the actions
will be auto enabled / disabled when switching view-spaces according to the
available history for that view-space

Signed-off-by: Waqar Ahmed <waqar.17a@gmail.com>