Allow opening results in new doc + use full doc text if no selection
This is a proposal to add the following two features to textfilter:
- Allow opening results in a new document
- If there is no selection, use full document text

This allows you to run grep / rg like tools to filter text in the
document quickly and view the results. One can do a selection to achieve
the same but the major issue with selection is that, it is very slow if
your document is big.