Do not use a half-created object
Registering the InlineNoteProvider in the constructor triggered a scan
of inline notes while the parent object (m_pluginView) was not yet fully
constructed. The inline note tried to accessed data in the
half-constructed parent and boom...

The registration in the constructor was not needed as the viewChanged
signal is anyways emitted after the plugin is constructed.