Use colors from KColorScheme
Instead of hard-coding colors, use colors from KColorScheme.

- Colors for showing line changes are still hard-coded. Not sure what to
use for them. If I use the KColorScheme colors for these, then you there
is no difference between 'M' / +12 -5 colors. This may be okay for some,
and for others not. I personally like having a little bit of difference
in colors. So, if you have any suggestions to improve this, do let me know

Signed-off-by: Waqar Ahmed <waqar.17a@gmail.com>