Tell user what to do instead of using sudo or kdesu
This warning shown when running Kate or KWrite using sudo or kdesu
is not very helpful, and leaves the user with the impression that
Kate and KWrite have no functionality for editing documents requiring
elevated privileges, when in fact the opposite is true. But we have
not told the user what to do instead of what they were trying to do,
so they do not know about it.

So let's tell the user what to do instead. :)