LSP: Improve tooltip
Before the lsp tooltip would ever be visible for 1 second when not in manual mode.

This makes the tooltip visible as long as the mouse has not moved, or hovers the tooltip. When the mouse moves there is a 100 delay before the tooltip is hidden to let the user hover over the tooltip if he wants to select text there.
If a tooltip is hovered when the mouse leaves it, the tooltip closes immediately.
There is only one Lsp tooltip ever shown.