Make collapsed sidebars resizable by mouse again
bccc7d350e6f81449acb7aec1983474f37c03685 hides plugin views when their
sidebar is collapsed which made the sidebar unresizable.

This commit adds a placeholder (an empty QLabel) which is shown
when the sidebar is collapsed so that it will still be resizable.
The plugin views are properly shown again after the sidebar is
manually expanded using the mouse.