lspclient: Enable copying diagnostics
This adds a custom menu entry to the diagnostics' context menu, which
simply copies the diagnostic message to the clipboard.

The targeted use cases are to ease the "make mistake → search for
mistake → find (wrong) solution online" workflow, as well as being able
to copy wrong warnings when reporting bugs against a LSP server
implementation.