gitblame: Show commit details in a file treeview
This change is a proposal to replace viewing the commit diff in a
tooltip with a treeview. Currently, you click on a commit and it opens
up a tooltip with the whole commit diff. This makes it a little bit
harder to use if your commit touched a lot of files or was large in
general.

With this change, once you click on a commit a temporary toolview  will be
created with list of all files from the commit. A user can then click on
a file to view the diff for that particular file in the commit.

This is a quickly done implementation so likely needs a lot of
refinements and tuning. Also, a lot of code is copied from the
CompareBranchView so we should probably find a way to remove this
duplication.