Show LspClient after positioning
Otherwise we are slightly wasteful on all platforms. On wayland it is
problematic as QtWayland does not currently support popup repositioning.

It looks like show() code was moved early so that we would have the size
ahead of positioning, but there is another way to achieve that. This is
the method that QMenu uses internally.

BUG: 443895