wayland fix window activation
Kate has it's own version of org.freedesktop.Application spec and
KDBusAddons handling. It needs to pass the activation token explicitly.

The launched (second) instance will have an environment variable
XDG_ACTIVATION_TOKEN set, we need to pass this value to the existing
application before it can set the window to activate.

Testing done: clicked on a file in dolphin, kate got focus