Introducing Colored Brackets plugin
This plugin allows to automatically color brackets in different colors.
This is helpful in stuff like:

```
if ((x && y) || (blah) || (foo)) { ... }
```

I have implemented in a way that only the current viewport's brackets are
checked. This allows us to achieve good results with very little work.
Also, another benefit is that the document length etc wouldn't affect
performance. Brackets are reparsed when you scroll or text is entered.