git-blame: Fix commit summary not available
This is a follow-up of my previous MR. The previous MR broke the
behaviour where we showed commit summary inline. This is now fixed.

To fix this, I had to switch to using `git blame -p` as the underlying
command to get blame info for a file. This I think is better because the
`p` switch is there for machine processing.