search: Allow filtering search results
Currently you can use search plugin to search for something in your
files/project. Often times there are a lot of results and you need to
filter in on the results. With this change you can filter on the results
quickly.

To do so, click the "Filter" tool button beside the "new tab" button. A
line edit should appear at the bottom of the search results. Results
will get filtered based on whatever you type in there.