Optimize project file filtering a bit
Typing into the filter line edit is too slow for the first couple of
chars. So when we have only one char, do simple string searching that
is much faster. For the rest, use simple fuzzy matching as we don't
use score here and all the extra work is not needed.