Only match filename by default in quickopen
Reasoning:
Firstly, I think it was just a very poor decision to match both if
filename matched. When filename got matched, it is unnecessary to match
the path and can lead to worse results if we do it. Consider files:

path/to/FileView.cpp
files/views/to/FileViewModel.cpp
...

First file match will happen, which most likely gives us the first file
for a pattern "fv". But then path matching which can ruin this result
and give some other irrelevant match a higher score leading to worse
results in quickopen

In kate, we work with files (from a user POV). User mostly sees filename
and will most likely be matching against file names, and rarely against
paths.

The other benefit that this gives is HUGELY better performance. This is
atleast 2x faster, so with large repos quick open will be much nicer to
use.

if one really wants to match a path, then specifying a pattern like:

src/fv

will do it. This is already supported, so it isn't something new.