Resize the Kate Session-Applet
Before this change there where only 4 fully visible items in the menu.
This makes visually searching very difficult as you need to scroll.

This change reduces the height of the menu items a tiny bit and
increases the height of the menu to match the height of the
Systemtray-menu. (Still smaller than the application menu/launcher)

(+ a binding loop fix)