Fix tab jumping with unlimited tabs
If we set the icon every time file is modified, it causes a lot of
layouting resulting in the tab bar getting scrolled.

With this change instead of setting an icon, we make the text italic to
show that this tab has modified content. This is more like other
editors.

BUG: 440585 435780