Urlbar: Collapse project base
With this change, the url bar will collapse the project base directory
whenever you have a file that belongs to a project in Kate. If you want
to navigate outside the project, clicking the "..." will expand the path
and allow you to navigate it.

Behind the scenes there are now two url bars, but for the user it will
appear as one.