Search&Replace: improve filtering behaviour and match counts
The filter now only checks the displayed part of the file path,
and excludes the base search path (which is the same for all
results anyway, so if it would match there, it would match for
all search results).

Also, the filtering for the filename is moved from the proxymodel
into the actual filter function in MatchModel. This way the displayed
match count is updated correctly (and it's less code).