Mouse forward/backward buttons configurable
Hi guys, love the project. I have been suffering from keyboard RSI for some time now and have been adapting to a more mouse oriented way of working to ease it. I am a big user of the LSP client and by issuing ctrl+click I can easily go to the definition of the symbol of interest but not so easily go back to where I was before. Consequently I added this configurabiliby to help me to go back and forward. This is my first contribution to the project so any feedback would be greatly appreciated.

Rather than being hardcoded to moving backwards and forwards between
open tabs, this commits enables the backward and forward buttons of
the mouse to be configurable to jump back and forwards on the history
stack as well.