Fix snippets not showing up in completion
If you have some text on a line:
	func()
and the cursor is in between the parenthesis and you try to invoke a
snippet, kate fails to do so because of what the completionRange
method was doing i.e., trying to select the full line range which imo
sounds not very smart.

It also leads to incorrect text replacement after execution of a snippet
as the range could be anything.

To test, create a snippet "qst" that completes to QStringLiteral("").
Then place your cursor in func() and type qst. Nothing shows up. Add
spaces between parenthesis, snippet will now show up but when you
execute it, it results in `tringLiteral("")`.

(cherry picked from commit b0b5da9c05dd49180b7fc44d3ae690c3cc604b57)