Fix startHostProcess() for KProcess < KF5.97
Since startHostProcess takes a QProcess and KProcess was modified to
behave more like a QProcess in 5.97, the #else needs to execute the
start() function of KProcess and not QPorcess. -> try to qobject_cast
the passed process to KProcess and use that if possible.


(cherry picked from commit 2403075c357ac60a46bd92a663dc83a256632749)