Replace Make Persistent with something better
These feature was the only way to have more than one tool in use at the same
time in one side bar. As long as you never clicked on some tab to switch some
tool was all fine. But when, you ended quickly in a situation where you desired
layout was damaged.

With this patch you can group tools in tab bars and and nicely switch between
them without to break your layout.

- Wrap KMultiTabBar in new class MultiTabBar
  It pointed out, that the KMultiTabBar only render properly when they
  is placed inside a QLayout. The new class do now much more than only to ensure
  the render works, so the name is not the best
- Use a QStackedWidget to simplify splitter sync tabs/tools
- Remove setSplitter(...) function. Not needed/used. Perhaps in the past or for
  future plans to move a Sidebar around?
- Add/Remove m_resizePlaceholder dynamically. A permanent existent extra widget
  cause trouble with splitter sync
- Don't use ownSpilt as parent for new ToolView. This should avoid unneeded
  Qt bookkeeping
- Fix check in handleCollapse() if moved splitter is of interest. The obscured
  math was wrong (or outdated). Add m_ownSplitIndex to avoid redundant calls
- Remove obsolete stuff and rename some other to improve readability