Remove close button from default toolbar
It's redundant:
- If the current document is the only one, you can use the window's
  close button to close the window.
- If the current document is one of many, you can use the close button
  on that document's tab to close it.

The only non-redundant workflow that this toolbar button supports is
the workflow of closing the last open document to get back to the
welcome screen again, which seems like quite a weird niche thing to do.
Accordingly, we can reduce some visual clutter on the toolbar by
removing this button and the separators around it.