project: Add ability to exclude folders
Sometimes we have folder around in the project that don't need to be
shown in the project tree. For git projects, you can just gitignore them
but for non-git folder based projects there is no way to do this.

With this commit, you can now specify a "exclude_folder_patterns" in the
.kateproject file like:

{
   "exclude_patterns": ["node_modules", "blah"]
}

and any file/folder that matches this pattern is ignored