Allow detaching search tab to mainWindow
When there are a lot of results and you have to go through them all,
a bigger widget can be very helpful.

Once detached, it will be a _read only_ widget i.e., you can only view
the search results and click on a result to jump to the document.