KateFileTree: fix renaming files if dest already exists
The old code was using a blocking exec() call, which meant that if you try
to rename a file and get the "file already exists" dialog from KIO, the
dialog and Kate are both blocked.

This was fixed some time ago in KateFileActions, and now that it's part of
a kateprivate lib, we can use it in the addons.