Introduce ByteArraySplitter
Small helper class to split QByteArray on a char. For now splitting only
on char is supported but can be extended to split on a string if needed.

The main reason to include this is to avoid indexOf based iteration that
I am using in many places which makes the code difficult to read. With
this, we can use normal c++ range for to nicely iterate over splitted
parts of the string.