Make LSP context menu actually contextual
Currently the LSP context menu is a submenu of the main context menu.
However, some LSP actions are arguably important enough that they
deserve to be in a most prominent spot in the main context menu.

This ensures LSP context menu actions are only shown when there is a
server.