KateTabBar: Add document icon to drag pixmap
Makes it more consistent with the item the drag originates from.
Also, set `State_Enabled` so the color set is correct.

Adding the close button is a lot more involved as it is an actual
`QPushButton` on `QTabBar` side but adding the icon is easy.


(cherry picked from commit a69e201471a05e1944eee11e5e20c1a84ce7030e)