Introduce CommitView
Moved out of git blame plugin. It can git show a commit in a tree view.
Moved out so that it can be used elsewhere. For now I have added an
action which takes a commit hash and shows that commit hash. The action
is accessible via GitWidget hamburger menu and is named 'Open Commit...'