Added Commands To Copy and Cut Searched Lines
The new commands ("Copy Searched Lines" and "Cut Searched Lines") are
hosted respectively in the "Edit->Copy Variants" and
"Edit->Cut Variants" sub menus. The existing command "Copy Location"
has been moved into the "Copy Variants" sub menu.

The new sub menus have been created with the idea that there should
be a reasonable amount of actions registered in them.  There are
many examples of copy commands in similar text editors, and these
"Variant" menus are a great way to surface this functionality to
new users looking for Kate functionality via the top menu.

An example use of Cut Searched Lines is to allow the user to quickly
pare down and analyse a complicated text file to progressively remove
more and more "uninteresting" and similar text. After a process of
several of these "a); search; b) cut" actions - the user will arrive
at the remains of the text file which should hold any
"unexpected/interesting" content for further research.

Optionally, during the process, the user may paste into a scratch
document to review these similar lines in bulk (at a glance) to
confirm that nothing interesting is present before moving on with
the next steps.  This could be for gaining an understanding of a
dense generated html file or an application source code file which
contains a lot of copy/pasted blocks.  It could also be used for a
kde menu ".rc" file to find any interesting menu attributes in use.

Copy Searched Lines may be used to capture data for a "table of
contents" or an index.  If the content uses a convention, then a
regular expression can be used to specify lines that represent
headings for capture with this command.

There are many situations for which this kind of processing is fast,
and produces understandable results, aiding the user quickly. You can
possibly think of other "needle(s) in the haystack" type of analysis,
where this type of processing can be of great benefit.