Show a simpler dialog when closing a window with only one changed file
With this, the KateSaveModifiedDialog distinguishes if only one file has
been changed, or multiple ones. If the former is the case, it looks
similar to a QMessageBox::warning dialog only asking to save or discard
that very file (or to cancel of course).

If multiple files have been changed, the dialog looks like before.

This way, we get a more straightforward UI, as the QTreeWidget shown by
default is pointless if we only talk about one file. Also, we get the
behavior of good ole KWrite back a bit more ;-)