In kate-ctags, go to declaration if no definition found
For some languages, what would be commonly seen as the definition of a
function does not get a ctags kind corresponding to a 'definition' type as
listed in gotoDefinition(). For example, Python class member functions
are of kind 'm', which kate-ctags identifies as a 'declaration' type in
gotoDeclaration(). The result is that right now, using the 'Go to
Definition' does not give any results while searching for a Python class
member function.
To overcome this while not jumping into the trouble of having a
per-language type list, this is a proposal to simply try to find a
'declaration' if no 'definition' was found while looking for one.