Fix clash between konsole's search shortcut and ours
Ctrl+Shift+F is also a shortcut exposed by konsole and we also use it
in search plugin. The one in search plugin is global and thus it clashes
with the one in konsole if konsole is active.

Perhaps there is a cleaner way to fix this, but I haven't found it yet.