Fix katebuild error/warning colors
Sets new colors for error and warning, fixes the too strict regexp and adds case-insensitive checking.

I was going to use kcolorscheme for the colors, but there is no warning and error colors.

Closest we have are negative (error) and neutral (warning) but I am not sure if those are supposed to be used like that.

If it's fine to use negative and neutral colors for errors and warnings, I can quickly change it.