Improve shortcut context / handling of search plugin
By default, KXMLGUI will add the widget of the guiFactory to the
associated widgets of the actionCollection. In this case, the widget
is MainWindow and thus the shortcut becomes active in the whole window.
This leads to shortcut conflicts with other widgets that are using same
shortcuts e.g., Konsole.

Instead, associate the ViewManager to search plugin's actionCollection
so that actions can be activated when a view has focus.