Introduce new terminal that works on windows/mac/linux
This change adds a new terminal based on QTermWidget/Konsole/PtyQt
that works on windows, linux and mac(untested).

QTermWidget by default doesn't support Windows, but I modified it by
reusing some parts from the Konsole codebase and by using PtyQt as a pty
backend. So, it is a little bit of a frankestein class.

The main reason for choosing QTermWidget widget is that
- its Qt and C++
- its basically Konsole, just a bit older so will be easy to reuse it in
  places where our existing Konsole plugin works

This change adds it as a plugin but I would propsoe (in the future) to
remove the plugin and just ifdef add it in our existing plugins which
use terminal.

If accepted, I would propose to rebase this on Konsole's code e.g., we
can upgrade the VT parser, konsole guys recently rewrote it and improved
it a lot. Moreover, this can serve as a reference point for anyone who
wants to port konsole to windows