[gdbplugin] be aware of multiple breakpoints in line
Since the debugger can move the requested break-point to the
nearest line with a statement, there can be more than one line-based
break-point at the same location.

This change is to ensure all the replicas are removed at once, because
the request is asking for line and not break-point ids, and to prevent
the function `toggleBreakpoint' to be confused.

A document should keep a line highlighted while there are at least one
breakpoint in the line.