Add QML language server configuration
This adds the new QML language server to the default LSP server
configuration.

The QML language server is in a usable state since Qt 6.4.1 and the Qt
company intends to add more features to it, so it makes sense to add it
as the default LSP handling QML files now.

An alternative LSP implementation is available at:
https://invent.kde.org/sdk/qml-lsp

However given that qmlls has the backing of the Qt company, it is fair
to say that qmlls will probably outlive qml-lsp.