Intoduce KateTextHintManager
KateTextHintManager provides centralized async interface to show
text hints. It uses KTextEditor::TextHintInterface::textHint() to know
when a text hint it is requested. It then sends a request to all
providers to create text hint and send it back when the text hint is
ready. This allows us to show information from multiple sources inside
one tooltip and avoids that everyone has to implement a provider and
manage it from scratch.

BUG: 467322