Add colors and clickable links for info/warning/errors in build output
Normally I'm only interested in the first warning/error in the build output and it is nice to have the context close by.

This MR adds background to the errors and lets the user click the error to jump to the line of the error. All in the raw output of the build plugin.
- The scrolling stops when an error/warning/info occurs. and can continue if you scroll to the bottom.
- (Feeling a bit bad about it) does not jump to the diagnostics view on completion..
- Only updates the TextBrowser every 100ms to minimize the CPU usage for updating the output
- Remove the broken "side-panel" mode + fix margins

@akselmo & @waqar, what do you think?

P.S. the next/previous error through Diagnostics view still works (but it does not update the raw output scroll)