Debug buttons in debug toolview
The continue and start button is combined, since there's not much reason to have different buttons for the two. 

The toolbar is still there for those wanting more fine grained control + more features. But these are the basic debug features
that are always visible:
- Start/Continue
- Stop
- Jump in
- Jump over
- Jump out
- Run to cursor
- Restart

The buttons follow same actions as the toolbar buttons, so they're basically same thing in different place. They get disabled and enabled similar ways.