kateviewmanager: Show welcome page as for new windows when all tabs closed
When the option to show the welcome page for new windows is set, if all tabs are closed this will now show the welcome page rather than a new document.
If a welcome page is already open, it is now activated rather than doing nothing or opening another - also for the welcome page action.