lspInlayHints: Fix issues with servers not honoring requested range
Dart-analyzer always send inlay hints for the full document even if we
just asked it for one line. This slightly messes up our cache in some
scenarios.

Example: you explicitly specify the type of a variable. The
new inlayHints sent by the server will not have the inlayHint that
was showing the variable type because now we explicitly wrote it in the
code. However, our cache updates in a way that it will merge the
incoming hints into current ones and we will end up keeping the
inlayHint that is no longer valid