lsp: always honor server when using Completion.textEdit.newText
If server is specifying a different range start than what we have,
perfer the server always.

Only do this if Completion.textEdit.newText is the text we are going to
insert. With pure insertText things are generally simpler.

This fixes issues with dart-analyzer and typescript-language-server

BUG: 468130
(cherry picked from commit bb94925014f710ea89752949d5ab90aefafd2175)