KateMainWindow: Prevent hamburger button from being pushed into the overflow
Qt lays out toolbars left to right, so we can end up in a situation where the
main toolbar overflows and the hamburgerBar with the single hamburger menu
button also gets an overflow.

This makes the menu button hard to access and two carets next to each other
looking terrible.