Fix crash on viewspace close
When we close a viewspace, we delete all the widgets implicitly (they
are part of QStackWidget which controls their lifetime). Moving around
widgets doesn't make much sense so instead, just skip them and let them
be killed.

Added a related test case.

As a drive-by, I discovered that there was still a check in place which
would prompt you and make sure you wanted to close the viewspace if the
viewspace has widgets. This is kind of wrong as we don't need to do this.
We should call the widget's shouldClose method and the widget should
prompt the user if necessary.


(cherry picked from commit 315a92d6b0dc4c2f4b8c5ab3169ebc3e8e6cbc73)