Add LSP support for window/showMessageRequest
Add a signal showMessageRequest to lspclientserver, and an identical one to
lspclientservermanager that retransmits the initial ones. Connect this to
a slot in LSPClientPluginViewImpl that shows a message box (a modified
QMessageBox that allows for the window close button to close the box).

Also fix parseMessage for LSPShowMessageParams (as showMessageRequest
duplicates showMessage's params) so that it does not always give the default
type.